Use Long quantity in ventas model

+++ src/main/java/microservice/ventas/dto/DescuentoStockRequest.java
@@ -11,6 +11,6 @@ public class DescuentoStockRequest {
 
     private Long idProducto;
     private Long idSucursal;
-    private Integer cantidad;
+    private Long cantidad;
     private String motivo;
 }
